fix(sql-editor): default data source
