+++ frontend/src/views/sql-editor/AsidePanel/SchemaPane/actions.tsx
@@ -1,3 +1,4 @@
+import { head } from "lodash-es";
 import {
   CodeIcon,
   CopyIcon,
@@ -13,11 +14,7 @@ import { useExecuteSQL } from "@/composables/useExecuteSQL";
 import { t } from "@/plugins/i18n";
 import { PROJECT_V1_ROUTE_DATABASE_DETAIL } from "@/router/dashboard/projectV1";
 import { SQL_EDITOR_DATABASE_MODULE } from "@/router/sqlEditor";
-import {
-  pushNotification,
-  useAppFeature,
-  useSQLEditorTabStore,
-} from "@/store";
+import { pushNotification, useAppFeature, useSQLEditorTabStore } from "@/store";
 import {
   DEFAULT_SQL_EDITOR_TAB_MODE,
   type ComposedDatabase,
@@ -25,6 +22,7 @@ import {
   type Position,
 } from "@/types";
 import { Engine } from "@/types/proto/v1/common";
+import { DataSource, DataSourceType } from "@/types/proto/v1/instance_service";
 import {
   defer,
   extractInstanceResourceName,
@@ -325,6 +323,7 @@ const runQuery = async (
       database: database.name,
       schema,
       table: tableOrViewName,
+      dataSourceId: getDefaultQueriableDataSourceOfDatabase(database).id,
     },
     mode: DEFAULT_SQL_EDITOR_TAB_MODE,
     worksheet: "",
@@ -390,4 +389,15 @@ export const selectAllFromTableOrView = async (
     return;
   }
   runQuery(db, schema, tableOrViewName, query);
+};
+
+const getDefaultQueriableDataSourceOfDatabase = (
+  database: ComposedDatabase
+) => {
+  const dataSources = database.instanceResource.dataSources;
+  const readonlyDataSources = dataSources.filter(
+    (ds) => ds.type === DataSourceType.READ_ONLY
+  );
+  // First try to use readonly data source if available.
+  return (head(readonlyDataSources) || head(dataSources)) as DataSource;
 };
